# _15_appliance_store/api/store/v1/serviceOrder_router.pb.go
// Code generated by https://github.com/go-dev-frame/sponge, DO NOT EDIT.

package v1

import (
	"context"
	"errors"
	"strings"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"

	"github.com/go-dev-frame/sponge/pkg/errcode"
	"github.com/go-dev-frame/sponge/pkg/gin/middleware"
)

type ServiceOrderLogicer interface {
	Create(ctx context.Context, req *CreateServiceOrderRequest) (*CreateServiceOrderReply, error)
	DeleteByID(ctx context.Context, req *DeleteServiceOrderByIDRequest) (*DeleteServiceOrderByIDReply, error)
	UpdateByID(ctx context.Context, req *UpdateServiceOrderByIDRequest) (*UpdateServiceOrderByIDReply, error)
	GetByID(ctx context.Context, req *GetServiceOrderByIDRequest) (*GetServiceOrderByIDReply, error)
	List(ctx context.Context, req *ListServiceOrderRequest) (*ListServiceOrderReply, error)
}

type ServiceOrderOption func(*serviceOrderOptions)

type serviceOrderOptions struct {
	isFromRPC  bool
	responser  errcode.Responser
	zapLog     *zap.Logger
	httpErrors []*errcode.Error
	rpcStatus  []*errcode.RPCStatus
	wrapCtxFn  func(c *gin.Context) context.Context
}

func (o *serviceOrderOptions) apply(opts ...ServiceOrderOption) {
	for _, opt := range opts {
		opt(o)
	}
}

func WithServiceOrderHTTPResponse() ServiceOrderOption {
	return func(o *serviceOrderOptions) {
		o.isFromRPC = false
	}
}

func WithServiceOrderRPCResponse() ServiceOrderOption {
	return func(o *serviceOrderOptions) {
		o.isFromRPC = true
	}
}

func WithServiceOrderResponser(responser errcode.Responser) ServiceOrderOption {
	return func(o *serviceOrderOptions) {
		o.responser = responser
	}
}

func WithServiceOrderLogger(zapLog *zap.Logger) ServiceOrderOption {
	return func(o *serviceOrderOptions) {
		o.zapLog = zapLog
	}
}

func WithServiceOrderErrorToHTTPCode(e ...*errcode.Error) ServiceOrderOption {
	return func(o *serviceOrderOptions) {
		o.httpErrors = e
	}
}

func WithServiceOrderRPCStatusToHTTPCode(s ...*errcode.RPCStatus) ServiceOrderOption {
	return func(o *serviceOrderOptions) {
		o.rpcStatus = s
	}
}

func WithServiceOrderWrapCtx(wrapCtxFn func(c *gin.Context) context.Context) ServiceOrderOption {
	return func(o *serviceOrderOptions) {
		o.wrapCtxFn = wrapCtxFn
	}
}

func RegisterServiceOrderRouter(
	iRouter gin.IRouter,
	groupPathMiddlewares map[string][]gin.HandlerFunc,
	singlePathMiddlewares map[string][]gin.HandlerFunc,
	iLogic ServiceOrderLogicer,
	opts ...ServiceOrderOption) {

	o := &serviceOrderOptions{}
	o.apply(opts...)

	if o.responser == nil {
		o.responser = errcode.NewResponser(o.isFromRPC, o.httpErrors, o.rpcStatus)
	}
	if o.zapLog == nil {
		o.zapLog, _ = zap.NewProduction()
	}

	r := &serviceOrderRouter{
		iRouter:               iRouter,
		groupPathMiddlewares:  groupPathMiddlewares,
		singlePathMiddlewares: singlePathMiddlewares,
		iLogic:                iLogic,
		iResponse:             o.responser,
		zapLog:                o.zapLog,
		wrapCtxFn:             o.wrapCtxFn,
	}
	r.register()
}

type serviceOrderRouter struct {
	iRouter               gin.IRouter
	groupPathMiddlewares  map[string][]gin.HandlerFunc
	singlePathMiddlewares map[string][]gin.HandlerFunc
	iLogic                ServiceOrderLogicer
	iResponse             errcode.Responser
	zapLog                *zap.Logger
	wrapCtxFn             func(c *gin.Context) context.Context
}

func (r *serviceOrderRouter) register() {
	r.iRouter.Handle("POST", "/api/v1/serviceOrder", r.withMiddleware("POST", "/api/v1/serviceOrder", r.Create_44)...)
	r.iRouter.Handle("DELETE", "/api/v1/serviceOrder/:id", r.withMiddleware("DELETE", "/api/v1/serviceOrder/:id", r.DeleteByID_42)...)
	r.iRouter.Handle("PUT", "/api/v1/serviceOrder/:id", r.withMiddleware("PUT", "/api/v1/serviceOrder/:id", r.UpdateByID_42)...)
	r.iRouter.Handle("GET", "/api/v1/serviceOrder/:id", r.withMiddleware("GET", "/api/v1/serviceOrder/:id", r.GetByID_42)...)
	r.iRouter.Handle("POST", "/api/v1/serviceOrder/list", r.withMiddleware("POST", "/api/v1/serviceOrder/list", r.List_44)...)

}

func (r *serviceOrderRouter) withMiddleware(method string, path string, fn gin.HandlerFunc) []gin.HandlerFunc {
	handlerFns := []gin.HandlerFunc{}

	// determine if a route group is hit or miss, left prefix rule
	for groupPath, fns := range r.groupPathMiddlewares {
		if groupPath == "" || groupPath == "/" {
			handlerFns = append(handlerFns, fns...)
			continue
		}
		size := len(groupPath)
		if len(path) < size {
			continue
		}
		if groupPath == path[:size] {
			handlerFns = append(handlerFns, fns...)
		}
	}

	// determine if a single route has been hit
	key := strings.ToUpper(method) + "->" + path
	if fns, ok := r.singlePathMiddlewares[key]; ok {
		handlerFns = append(handlerFns, fns...)
	}

	return append(handlerFns, fn)
}

func (r *serviceOrderRouter) Create_44(c *gin.Context) {
	req := &CreateServiceOrderRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.Create(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *serviceOrderRouter) DeleteByID_42(c *gin.Context) {
	req := &DeleteServiceOrderByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeleteByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *serviceOrderRouter) UpdateByID_42(c *gin.Context) {
	req := &UpdateServiceOrderByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UpdateByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *serviceOrderRouter) GetByID_42(c *gin.Context) {
	req := &GetServiceOrderByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.GetByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *serviceOrderRouter) List_44(c *gin.Context) {
	req := &ListServiceOrderRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.List(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}