# _15_appliance_store/docs/apis.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "store api docs",
    "version": "2.0"
  },
...
  "host": "localhost:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/afterSales/service-orders": {
      "post": {
        "summary": "创建维修工单\n实现逻辑：\n1. 验证客户购买记录有效性\n2. 检查产品是否在保修期内\n3. 分配就近服务网点\n4. 生成带条形码的维修工单",
        "operationId": "afterSales_CreateServiceOrderWithAssignment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateServiceOrderWithAssignmentReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateServiceOrderWithAssignmentRequest"
            }
          }
        ],
        "tags": [
          "afterSales"
        ]
      }
    },
...
    "/api/v1/serviceOrder": {
      "post": {
        "summary": "create serviceOrder",
        "operationId": "serviceOrder_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateServiceOrderReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateServiceOrderRequest"
            }
          }
        ],
        "tags": [
          "serviceOrder"
        ]
      }
    },
    "/api/v1/serviceOrder/list": {
      "post": {
        "summary": "list of serviceOrder by query parameters",
        "operationId": "serviceOrder_List",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListServiceOrderReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ListServiceOrderRequest"
            }
          }
        ],
        "tags": [
          "serviceOrder"
        ]
      }
    },
    "/api/v1/serviceOrder/{id}": {
      "get": {
        "summary": "get serviceOrder by id",
        "operationId": "serviceOrder_GetByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetServiceOrderByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "serviceOrder"
        ]
      },
      "delete": {
        "summary": "delete serviceOrder by id",
        "operationId": "serviceOrder_DeleteByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteServiceOrderByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "serviceOrder"
        ]
      },
      "put": {
        "summary": "update serviceOrder by id",
        "operationId": "serviceOrder_UpdateByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateServiceOrderByIDReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "customerID": {
                  "type": "integer",
                  "format": "int64"
                },
                "productID": {
                  "type": "integer",
                  "format": "int64"
                },
                "faultDesc": {
                  "type": "string"
                },
                "status": {
                  "type": "integer",
                  "format": "int32"
                },
                "technicianID": {
                  "type": "integer",
                  "format": "int64"
                },
                "completionTime": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "serviceOrder"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "typesColumn": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "exp": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "logic": {
          "type": "string"
        }
      }
    },
    "typesParams": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        },
        "sort": {
          "type": "string"
        },
        "columns": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/typesColumn"
          }
        }
      }
    },
...
    "v1CreateServiceOrderReply": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "v1CreateServiceOrderRequest": {
      "type": "object",
      "properties": {
        "customerID": {
          "type": "integer",
          "format": "int64"
        },
        "productID": {
          "type": "integer",
          "format": "int64"
        },
        "faultDesc": {
          "type": "string"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "technicianID": {
          "type": "integer",
          "format": "int64"
        },
        "completionTime": {
          "type": "string"
        }
      }
    },
    "v1CreateServiceOrderWithAssignmentReply": {
      "type": "object",
      "properties": {
        "serviceOrderID": {
          "type": "string"
        },
        "assignedStore": {
          "type": "string"
        },
        "qrCodeURL": {
          "type": "string"
        }
      }
    },
    "v1CreateServiceOrderWithAssignmentRequest": {
      "type": "object",
      "properties": {
        "customerID": {
          "type": "integer",
          "format": "int64"
        },
        "productID": {
          "type": "integer",
          "format": "int64"
        },
        "faultDescription": {
          "type": "string"
        },
        "purchaseProof": {
          "type": "string"
        }
      }
    },
...
    "v1DeleteServiceOrderByIDReply": {
      "type": "object"
    },
...
    "v1GetServiceOrderByIDReply": {
      "type": "object",
      "properties": {
        "serviceOrder": {
          "$ref": "#/definitions/v1ServiceOrder"
        }
      }
    },
...
    "v1ListServiceOrderReply": {
      "type": "object",
      "properties": {
        "total": {
          "type": "integer",
          "format": "int64"
        },
        "serviceOrders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1ServiceOrder"
          }
        }
      }
    },
    "v1ListServiceOrderRequest": {
      "type": "object",
      "properties": {
        "params": {
          "$ref": "#/definitions/typesParams"
        }
      }
    },
...
    "v1ServiceOrder": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "customerID": {
          "type": "integer",
          "format": "int64"
        },
        "productID": {
          "type": "integer",
          "format": "int64"
        },
        "faultDesc": {
          "type": "string"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "technicianID": {
          "type": "integer",
          "format": "int64"
        },
        "completionTime": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
...
    "v1UpdateServiceOrderByIDReply": {
      "type": "object"
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "description": "Type Bearer your-jwt-token to Value",
      "name": "Authorization",
      "in": "header"
    }
  }
}